feat: fetch responsible editor
# Motivation

To be able to create a corpus of cleaned web
archived material, some additional work needs to
be done. Specifically, only text that have a
responsible editor can be used.

# Changes

This commit adds functionality to fetch the full
text of the database entries with responsible
editors.

# Interface

This is how the interface looks like

```
python main.py --help
usage: main.py [-h] --filter-yaml-file FILTER_YAML_FILE
--hostname HOSTNAME --port PORT --database DATABASE
--user USER --password PASSWORD --output-dir OUTPUT_DIR

options:
  -h, --help            show this help message and exit
  --filter-yaml-file FILTER_YAML_FILE
                        Path to the filter yaml file
  --hostname HOSTNAME   Hostname of the database
  --port PORT           Port of the database
  --database DATABASE   Database name
  --user USER           Username of the database
  --password PASSWORD   Password of the database
  --output-dir OUTPUT_DIR
                        Path to the output dir
```

# How to run

See the readme for instructions on how to run the
script.

# Future work

Avoid passing the password as plaintext, which is
not great with regards to security.