feat: add metadata to `yaml` file
# Motivation

When creating a corpus with the soon to be
tokenized text, a few other details are required.
Such as `URL`, `date`, etc

# Metadata

The fields added by this commit are:
- harvest-timestamp
- URL
- dhlab-id

`dhlab-id` is an optional field to help produce a
correct ID for the Digital Humanities Lab at the
National Library of Norway.

This field can be disabled by using the argument
`--disable-dhlab-id`.

# Future work

Replace the full text with a tokenized version of
the full text.